requirements.txt
asgiref==3.5.2
cairocffi==1.3.0
CairoSVG==2.5.2
cffi==1.15.0
cssselect2==0.6.0
defusedxml==0.7.1
Django==4.0.5
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
lxml==4.9.0
numpy==1.22.4
Pillow==9.1.1
pycairo==1.21.0
pycparser==2.21
pylunasvg==0.0.2
pytz==2022.1
reportlab==3.6.9
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tzdata==2022.1
webencodings==0.5.1
